requirements.txt
flask>=3.0.0
playwright>=1.40.0
python-pptx>=0.6.21
img2pdf>=0.5.0
